[refactor] DTO들의 패키지 위치를 변경하고, 네이밍에 통일성을 주려 하였습니다.
