+++ be/src/main/java/kr/codesquad/sidedish/controller/RequestProduct.java
@@ -1,4 +1,4 @@
-package kr.codesquad.sidedish.dto;
+package kr.codesquad.sidedish.controller;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
